refactor: delete AbstractChangeConfigurationParams
reason: it was created when we had version specific enums and therefore
ChangeConfigurationParams. with the current impl, it has no use and
created only confusion.
